requirements.txt
Scrapy~=2.7.1
itemadapter~=0.7.0
SQLAlchemy~=1.4.43
alembic~=1.8.1
python-dotenv~=0.21.0
fastapi~=0.86.0
uvicorn~=0.19.0
psycopg2-binary==2.9.3